pyproject.toml
[project]
name = "recipeprinter"
version = "0.1dev"
dependencies = [
    "psycopg2",
    "flask",
    "python-dotenv"
]